refactor(lessonstable): 规范了lessonstable下的组件命名，以及其他的规范问题
