+++ src/pages/lessonstable/_components/lesson-grid/dark.scss
