+++ src/pages/lessonstable/_components/lesson-grid/light.scss
